old-llm-assistant-chatbot-tuto/requirements.txt
llama-cpp-python
outlines
transformers
huggingface_hub
duckduckgo-search
chromadb
llama-index-core
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
llama-index-readers-web
docker
